lesson 7 make folder Commands
