+++ TodoList/TodoList/Commands/ComandParses.cs
@@ -1,9 +1,5 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-namespace TodoList
+namespace TodoList.Commands
 {
     static class CommandParser
     {
