+++ TodoList/TodoList/Commands/DoneComands.cs
@@ -1,28 +1,22 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-namespace TodoList
+namespace TodoList.Commands
 {
-    internal class UpdateCommand : ICommand
+    internal class DoneCommand : ICommand
     {
         public TodoList List { get; set; }
         public int Index { get; set; }
-        public string NewText { get; set; }
-        public UpdateCommand(TodoList list, int index, string newText)
+        public DoneCommand(TodoList list, int index)
         {
             List = list;
             Index = index;
-            NewText = newText;
         }
         public void Execute()
         {
             var item = List.GetItem(Index);
             if (item != null)
             {
-                item.UpdateText(NewText);
-                Console.WriteLine($"Задача {Index} обновлена.");
+                item.MarkDone();
+                Console.WriteLine($"Задача {Index} отмечена как выполненная.");
             }
             else
             {
