+++ TodoList/TodoList/Commands/ICommand.cs
@@ -1,0 +1,8 @@
+﻿using System;
+namespace TodoList.Commands
+{
+    internal interface ICommand
+    {
+        void Execute();
+    }
+}
