+++ TodoList/TodoList/Commands/UpdateComand.cs
@@ -1,26 +1,24 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-namespace TodoList
+namespace TodoList.Commands
 {
-    internal class DoneCommand : ICommand
+    internal class UpdateCommand : ICommand
     {
         public TodoList List { get; set; }
         public int Index { get; set; }
-        public DoneCommand(TodoList list, int index)
+        public string NewText { get; set; }
+        public UpdateCommand(TodoList list, int index, string newText)
         {
             List = list;
             Index = index;
+            NewText = newText;
         }
         public void Execute()
         {
             var item = List.GetItem(Index);
             if (item != null)
             {
-                item.MarkDone();
-                Console.WriteLine($"Задача {Index} отмечена как выполненная.");
+                item.UpdateText(NewText);
+                Console.WriteLine($"Задача {Index} обновлена.");
             }
             else
             {
