+++ TodoList/TodoList/Commands/ViewComand.cs
@@ -1,6 +1,5 @@
 ﻿using System;
-
-namespace TodoList
+namespace TodoList.Commands
 {
     internal class ViewCommand : ICommand
     {
