+++ TodoList/TodoList/TodoItem.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using TodoList.Commands;
+using System;
 
 namespace TodoList
 {
